+++ .copier-answers.yml
@@ -6,10 +6,10 @@ author_name: Doug Branton
 create_example_module: true
 custom_install: true
 include_notebooks: true
-module_name: lsstseries
+module_name: tape
 mypy_type_checking: none
 preferred_linter: black
 project_license: MIT
-project_name: lsstseries
+project_name: tape
 use_gitlfs: none
 use_isort: false
