+++ docs/notebooks.rst
@@ -3,8 +3,8 @@ Notebooks
 
 .. toctree::
 
-    Working with the lsstseries Ensemble object <notebooks/working_with_the_ensemble>
-    Working with the lsstseries Timeseries object <notebooks/working_with_the_timeseries>
+    Working with the TAPE Ensemble object <notebooks/working_with_the_ensemble>
+    Working with the TAPE Timeseries object <notebooks/working_with_the_timeseries>
     Scaling to Large Data Volume <notebooks/scaling_to_large_data>
     Working with Structure Function <notebooks/working_with_structure_function>
     Binning Sources in the Ensemble <notebooks/binning_slowly_changing_sources>
