+++ src/tape/analysis/light_curve.py
