+++ src/tape/analysis/stetsonj.py
