+++ src/tape/analysis/structure_function/__init__.py
