+++ src/tape/analysis/structure_function/base_argument_container.py
