+++ src/tape/analysis/structure_function/base_calculator.py
