+++ src/tape/analysis/structure_function/bauer_2009a/calculator.py
