+++ src/tape/analysis/structure_function/bauer_2009b/calculator.py
