+++ src/tape/analysis/structure_function/calculator_registrar.py
