+++ src/tape/analysis/structure_function/macleod_2012/calculator.py
