+++ src/tape/analysis/structure_function/sf_light_curve.py
