+++ src/tape/analysis/structurefunction2.py
@@ -3,8 +3,8 @@ from typing import List
 import numpy as np
 import pandas as pd
 
-from lsstseries.analysis.structure_function import SF_METHODS
-from lsstseries.analysis.structure_function.sf_light_curve import StructureFunctionLightCurve
+from tape.analysis.structure_function import SF_METHODS
+from tape.analysis.structure_function.sf_light_curve import StructureFunctionLightCurve
 
 
 def calc_sf2(time, flux, err=None, band=None, lc_id=None, sf_method="basic", argument_container=None):
