+++ src/tape/timeseries.py
