+++ tests/tape_tests/conftest.py
@@ -2,8 +2,8 @@
 import pytest
 from dask.distributed import Client
 
-from lsstseries import Ensemble
-from lsstseries.utils import ColumnMapper
+from tape import Ensemble
+from tape.utils import ColumnMapper
 
 
 @pytest.fixture(scope="package", name="dask_client")
@@ -20,8 +20,8 @@ def parquet_ensemble(dask_client):
     """Create an Ensemble from parquet data."""
     ens = Ensemble(client=dask_client)
     ens.from_parquet(
-        "tests/lsstseries_tests/data/source/test_source.parquet",
-        "tests/lsstseries_tests/data/object/test_object.parquet",
+        "tests/tape_tests/data/source/test_source.parquet",
+        "tests/tape_tests/data/object/test_object.parquet",
         id_col="ps1_objid",
         time_col="midPointTai",
         band_col="filterName",
@@ -38,7 +38,7 @@ def parquet_ensemble_from_source(dask_client):
     """Create an Ensemble from parquet data, with object file withheld."""
     ens = Ensemble(client=dask_client)
     ens.from_parquet(
-        "tests/lsstseries_tests/data/source/test_source.parquet",
+        "tests/tape_tests/data/source/test_source.parquet",
         id_col="ps1_objid",
         time_col="midPointTai",
         band_col="filterName",
@@ -63,7 +63,7 @@ def parquet_ensemble_with_column_mapper(dask_client):
         band_col="filterName",
     )
     ens.from_parquet(
-        "tests/lsstseries_tests/data/source/test_source.parquet",
+        "tests/tape_tests/data/source/test_source.parquet",
         column_mapper=colmap,
     )
 
@@ -78,7 +78,7 @@ def parquet_ensemble_with_known_column_mapper(dask_client):
 
     colmap = ColumnMapper().use_known_map("ZTF")
     ens.from_parquet(
-        "tests/lsstseries_tests/data/source/test_source.parquet",
+        "tests/tape_tests/data/source/test_source.parquet",
         column_mapper=colmap,
     )
 
@@ -91,7 +91,7 @@ def parquet_ensemble_from_hipscat(dask_client):
     """Create an Ensemble from a hipscat/hive-style directory."""
     ens = Ensemble(client=dask_client)
     ens.from_hipscat(
-        "tests/lsstseries_tests/data",
+        "tests/tape_tests/data",
         id_col="ps1_objid",
         time_col="midPointTai",
         band_col="filterName",
