+++ tests/tape_tests/structure_function_calculators/test_base_calculator.py
