+++ tests/tape_tests/structure_function_calculators/test_calculator_registrar.py
