+++ tests/tape_tests/structure_function_calculators/test_light_curve.py
