+++ tests/tape_tests/structure_function_calculators/test_sf_light_curve.py
