+++ tests/tape_tests/test_analysis.py
