+++ tests/tape_tests/test_timeseries.py
